pyproject.toml
[project]
name = "cash-flow"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "django-split-settings>=1.3.2",
    "drf-spectacular>=0.29.0",
    "django-stubs-ext>=5.2.7",
    "psycopg[binary]>=3.2.12",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.2.1",
    "django-cors-headers>=4.9.0",
    "celery>=5.6.0",
    "django-redis>=6.0.0",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-stubs[compatible-mypy]>=5.2.7",
    "djangorestframework-stubs[compatible-mypy]>=3.16.5",
    "pytest-django>=4.11.1",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]

[tool.ruff]
line-length = 85

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = ["celery.*", "debug_toolbar.toolbar.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "src.cash_flow.root.settings"
